+++ lmlib/include/lmlib/math_constants.h
@@ -2,5 +2,5 @@
 
 namespace lm
 {
-const float pi{3.141592653589793238462643};
+const float pi{(float)3.141592653589793238462643};
 }
